add support for RESTIC_PASSWORD env variable
The restic tool doesn't have a CLI parameter for this, but it allows
setting an env variable. We can check for that too and keep this as an
env variable only.

This way we keep the safety of knowing that all necessary parameters
(i.e. the repo and password options) are set.