prompt for password if none is given
This shows a password prompt, if neither the `--password-file` or
`--password-command` option is given **and** the `RESTIC_PASSWORD` env
variable is unset.

There's no way to hand over the password to restic as a CLI parameter,
so we use the `RESTIC_PASSWORD` env variable under the hood to set it
instead. The env variable is only valid for the restic process we call
internally, so the next time redu is run, it'll prompt for the password
again.